Hasher::write_u*() are expected keep the native endianness
So that write_u32(42) will produce different outputs on big and little
endian targets.

I couldn't even find that behavior properly documented. But this is
what the default trait implementations do.

This is absolutely awful. I don't even understand why they chose to do
that. Anyway, like a lot of things in Rust, we can just shut up and
comply.